% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{park_palette}
\alias{park_palette}
\title{A National Park palette generator}
\usage{
park_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{SmokyMountains}, \code{RockMountains},  \code{Yellowstone},
\code{Arches}, \code{ArticGates},  \code{MtMckinley}, \code{GeneralGrant},
\code{Hawaii},  \code{CraterLake} , \code{Saguaro} ,
\code{GrandTeton}, \code{BryceCanyon}, \code{MtRainier}, \code{Badlands},
\code{Redwoods}, \code{Everglades}, \code{Voyageurs}
@importFrom graphics rgb rect par image text}

\item{n}{Number of colors desired. Most palettes now only
have 5 to 8 colors. Most color
schemes are derived from \href{http://poppygall.com/blog/2011/07/25/color-inspiration-national-park-wpa-posters/}{Color Inspiration – National Park WPA Posters}.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from National Park posters and images.
}
\examples{
park_palette("SmokyMountains")
park_palette("Yellowstone", 3)
}
\keyword{colors}
