% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{park_palettes}
\alias{park_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 25.}
\usage{
park_palettes
}
\description{
Use \code{\link{park_palette}} to construct palettes of desired length.
}
\keyword{datasets}
